dash/requirements.txt
dash
dash-core-components
dash-html-components
opencv-python
dash-bootstrap-components
pillow
flask